% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrammer.R
\name{create_bigrams}
\alias{create_bigrams}
\title{Generate bigrams without stopwords from any text}
\usage{
create_bigrams(df, text)
}
\arguments{
\item{df}{A dataframe containing text to be processed}

\item{text}{The text variable to be processed}
}
\value{
A dataframe of bigrams
}
\description{
Generate bigrams without stopwords from any text
}
\examples{
url <- "https://fingertips.phe.org.uk/documents/PHDS\%20Guidance\%20-\%20Funnel\%20Plots.pdf"
doc <- readtext(url)
head(create_bigrams(doc, text))



}
